refactor: 도메인 내 연관된 객체끼리 패키지 세분화
