+++ src/main/java/controller/BlackjackController.java
@@ -5,13 +5,13 @@ import static constant.BlackjackConstant.DEALER_DRAW_BOUND;
 import static constant.BlackjackConstant.HIT_DRAW_COUNT;
 import static constant.BlackjackConstant.INIT_DRAW_COUNT;
 
-import domain.Card;
-import domain.CardDeck;
 import domain.CardResult;
-import domain.HandCards;
-import domain.Name;
-import domain.Participant;
-import domain.Participants;
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.HandCards;
+import domain.participant.Name;
+import domain.participant.Participant;
+import domain.participant.Participants;
 import java.util.ArrayList;
 import java.util.List;
 import service.BlackjackService;
