+++ src/main/java/domain/CardResult.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.card.Card;
 import java.util.List;
 
 public record CardResult(
