+++ src/main/java/domain/card/CardShufflerImpl.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import infra.CardShuffler;
