+++ src/test/java/domain/CardDeckTest.java
@@ -2,6 +2,10 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.CardRank;
+import domain.card.CardSuit;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
