+++ src/test/java/domain/CardShufflerImplTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.CardShufflerImpl;
 import infra.CardShuffler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
