+++ src/test/java/service/BlackjackServiceTest.java
@@ -4,14 +4,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 
 import constant.BlackjackConstant;
-import domain.Card;
-import domain.CardDeck;
-import domain.CardRank;
-import domain.CardSuit;
-import domain.HandCards;
-import domain.Name;
-import domain.Participant;
-import domain.Participants;
+import domain.card.Card;
+import domain.card.CardDeck;
+import domain.card.CardRank;
+import domain.card.CardSuit;
+import domain.card.HandCards;
+import domain.participant.Name;
+import domain.participant.Participant;
+import domain.participant.Participants;
 import infra.CardShuffler;
 import java.util.ArrayList;
 import java.util.List;
